+++ matmul/matmul.h
@@ -1,5 +1,5 @@
 // matmul.h
 #ifndef MATMUL_H
 #define MATMUL_H
-long* matmul(long *a, long *b, int m, int p, int n);
+float* matmul(float *a, float *b, int m, int p, int n);
 #endif
